finito ?

+++ srcs/ft_printf_specifier.c
@@ -6,7 +6,7 @@
 /*   By: amalangu <amalangu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/11 08:11:20 by amalangu          #+#    #+#             */
-/*   Updated: 2024/11/25 21:59:13 by amalangu         ###   ########.fr       */
+/*   Updated: 2025/01/14 14:04:57 by amalangu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,11 +23,11 @@ int	ft_printf_specifier(va_list args, const char *s)
 	else if (*s == 'd' || *s == 'i')
 		return (ft_printf_i(va_arg(args, int)));
 	else if (*s == 'u')
-		return (ft_printf_x(va_arg(args, unsigned int), *s));
+		return (ft_printf_x(va_arg(args, unsigned long), *s));
 	else if (*s == 'x')
-		return (ft_printf_x(va_arg(args, unsigned int), *s));
+		return (ft_printf_x(va_arg(args, unsigned long), *s));
 	else if (*s == 'X')
-		return (ft_printf_x(va_arg(args, unsigned int), *s));
+		return (ft_printf_x(va_arg(args, unsigned long), *s));
 	else if (*s == '%')
 		return (ft_putchar_int('%'));
 	return (-1);
